Servidor AlToque amb BBDD postgreSQL, peticions disponibles: - LOGIN - LOGOUT - USUARI_ADD - USUARI_DEL - EMPRESA_ADD - EMPRESA_DEL - EMPRESA_LLISTAR - PRODUCTE_ADD - PRODUCTE_DEL - PRODUCTE_LLISTAR

+++ src/main/java/com/altoque/altoque_server/dto/ProducteDto.java
@@ -8,12 +8,12 @@ public class ProducteDto {
     public Long id;
     public String nom;
     public String descripcio;
-    public Long preu;
+    public double preu;
     public String empresaCif;
 
     public ProducteDto() {}
 
-    public ProducteDto(Long id, String nom, String descripcio, Long preu, String empresaCif) {
+    public ProducteDto(Long id, String nom, String descripcio, Double preu, String empresaCif) {
         this.id = id;
         this.nom = nom;
         this.descripcio = descripcio;
